refactor(#730): un-nest enrich_artwork_results into context/top1/item/background submodules
Shape change #1 of the orchestrator decomposition: the two nested closures inside lookup/enrichment/__init__.py become module functions with explicit inputs. context.py carries the frozen EnrichmentContext (8 service handles + 5 request scalars — exactly what the closures captured; the plan's discogs_cache and warm_cache guesses are deliberately absent because only coordinator-inline code reads them). top1.py gets the former fetch_top1_release_details(ctx, items_with_artwork); item.py gets the former enrich_one plus _build_streaming_search_url (its only consumer), with the coordinator-derived top1 payload and LML#504 release-side flags as explicit keyword parameters; background.py gets the warm-cache machinery (_WARM_CACHE_CONCURRENCY, _warm_cache_semaphore, _background_tasks, _warm_bio_cache). __init__.py keeps enrich_artwork_results as a thin coordinator with an unchanged public signature. Bodies transplant verbatim modulo capture→ctx.field renames and dedent; no behavior change. Test patch targets retarget to the module that reads each name at call time (resolve_tracklist_via_musicbrainz ×15, apply_streaming_url_postprocess ×4, apple_music_lookup_timeout_s ×3, sentry_sdk.get_current_scope ×2 → lookup.enrichment.item; _background_tasks import → lookup.enrichment.background; asyncio.create_task ×4 stays on the coordinator). Also fixes two pre-existing stale pointers carried by the 6a move (artist_matches_item now cited at lookup/matching.py; the album-only-fallback note now described in module-path form) and rewords the orchestrator-as-actor prose in the transplanted docstrings; lookup/streaming_url_postprocess.py's cross-references follow the warm machinery to lookup.enrichment.background.
